requirements.txt
fastapi
uvicorn
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
scikit-learn
pandas==2.1.3
numpy
python-dateutil
pytz
APScheduler
sqlite3
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
